safari-600.7-branch/LayoutTests/platform/mac/accessibility/aria-liveregions-notifications-expected.txt

 new text element
This tests that ARIA live regions are sending out the correct notifications. We perform four operations (add, remove, change text, change alt tag), each one should trigger a live region notification

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS addedNotification is true
PASS successfullyParsed is true

TEST COMPLETE
